feat: split logic based on LE
