+++ packages/theme/src/ee/pages/dashboard/roadmaps/Index.vue
@@ -137,10 +137,12 @@ import { useSettingStore } from "../../../../store/settings";
 import { useUserStore } from "../../../../store/user";
 import {
   getAllRoadmaps,
+} from "../../../../modules/roadmaps";
+import {
   createRoadmap,
   sortRoadmap,
   deleteRoadmap
-} from "../../../../modules/roadmaps";
+} from "../../../modules/roadmaps";
 import { useCopyText } from "../../../../hooks";
 
 // components
